+++ src/mmu.rs
@@ -5,8 +5,8 @@ use gpu::GPU;
 use sound::Sound;
 use gbmode::{GbMode, GbSpeed};
 
-const WRAM_SIZE: uint = 0x8000;
-const ZRAM_SIZE: uint = 0x7F;
+const WRAM_SIZE: usize = 0x8000;
+const ZRAM_SIZE: usize = 0x7F;
 
 #[derive(PartialEq)]
 enum DMAType {
@@ -30,7 +30,7 @@ pub struct MMU<'a> {
 	hdma_src: u16,
 	hdma_dst: u16,
 	hdma_len: u8,
-	wrambank: uint,
+	wrambank: usize,
 	mbc: Box<::mbc::MBC+'static>,
 	gbmode: GbMode,
 	gbspeed: GbSpeed,
@@ -137,7 +137,7 @@ impl<'a> MMU<'a> {
 		self.gpu.gbmode = mode;
 	}
 
-	pub fn do_cycle(&mut self, cputicks: uint) -> uint {
+	pub fn do_cycle(&mut self, cputicks: u32) -> u32 {
 		let ticks = cputicks + self.perform_vramdma();
 
 		let gputicks = ticks /
@@ -165,8 +165,8 @@ impl<'a> MMU<'a> {
 			0x0000 ... 0x7FFF => self.mbc.readrom(address),
 			0x8000 ... 0x9FFF => self.gpu.rb(address),
 			0xA000 ... 0xBFFF => self.mbc.readram(address),
-			0xC000 ... 0xCFFF | 0xE000 ... 0xEFFF => self.wram[address as uint & 0x0FFF],
-			0xD000 ... 0xDFFF | 0xF000 ... 0xFDFF => self.wram[(self.wrambank * 0x1000) | address as uint & 0x0FFF],
+			0xC000 ... 0xCFFF | 0xE000 ... 0xEFFF => self.wram[address as usize & 0x0FFF],
+			0xD000 ... 0xDFFF | 0xF000 ... 0xFDFF => self.wram[(self.wrambank * 0x1000) | address as usize & 0x0FFF],
 			0xFE00 ... 0xFE9F => self.gpu.rb(address),
 			0xFF00 => self.keypad.rb(),
 			0xFF01 ... 0xFF02 => self.serial.rb(address),
@@ -178,7 +178,7 @@ impl<'a> MMU<'a> {
 			0xFF51 ... 0xFF55 => self.hdma_read(address),
 			0xFF68 ... 0xFF6B => self.gpu.rb(address),
 			0xFF70 => self.wrambank as u8,
-			0xFF80 ... 0xFFFE => self.zram[address as uint & 0x007F],
+			0xFF80 ... 0xFFFE => self.zram[address as usize & 0x007F],
 			0xFFFF => self.inte,
 			_ => { warn!("rb not implemented for {:X}", address); 0 },
 		}
@@ -193,8 +193,8 @@ impl<'a> MMU<'a> {
 			0x0000 ... 0x7FFF => self.mbc.writerom(address, value),
 			0x8000 ... 0x9FFF => self.gpu.wb(address, value),
 			0xA000 ... 0xBFFF => self.mbc.writeram(address, value),
-			0xC000 ... 0xCFFF | 0xE000 ... 0xEFFF => self.wram[address as uint & 0x0FFF] = value,
-			0xD000 ... 0xDFFF | 0xF000 ... 0xFDFF => self.wram[(self.wrambank * 0x1000) | (address as uint & 0x0FFF)] = value,
+			0xC000 ... 0xCFFF | 0xE000 ... 0xEFFF => self.wram[address as usize & 0x0FFF] = value,
+			0xD000 ... 0xDFFF | 0xF000 ... 0xFDFF => self.wram[(self.wrambank * 0x1000) | (address as usize & 0x0FFF)] = value,
 			0xFE00 ... 0xFE9F => self.gpu.wb(address, value),
 			0xFF00 => self.keypad.wb(value),
 			0xFF01 ... 0xFF02 => self.serial.wb(address, value),
@@ -206,8 +206,8 @@ impl<'a> MMU<'a> {
 			0xFF51 ... 0xFF55 => self.hdma_write(address, value),
 			0xFF68 ... 0xFF6B => self.gpu.wb(address, value),
 			0xFF0F => self.intf = value,
-			0xFF70 => { self.wrambank = match value & 0x7 { 0 => 1, n => n as uint }; },
-			0xFF80 ... 0xFFFE => self.zram[address as uint & 0x007F] = value,
+			0xFF70 => { self.wrambank = match value & 0x7 { 0 => 1, n => n as usize }; },
+			0xFF80 ... 0xFFFE => self.zram[address as usize & 0x007F] = value,
 			0xFFFF => self.inte = value,
 			_ => warn!("wb not implemented for {:X}", address),
 		};
@@ -240,7 +240,7 @@ impl<'a> MMU<'a> {
 
 	fn hdma_read(&self, a: u16) -> u8 {
 		match a {
-			0xFF51 ... 0xFF54 => { self.hdma[(a - 0xFF51) as uint] },
+			0xFF51 ... 0xFF54 => { self.hdma[(a - 0xFF51) as usize] },
 			0xFF55 => self.hdma_len | if self.hdma_status == DMAType::NoDMA { 0x80 } else { 0 },
 			_ => panic!("The address {:04X} should not be handled by hdma_read", a),
 		}
@@ -273,7 +273,7 @@ impl<'a> MMU<'a> {
 		};
 	}
 
-	fn perform_vramdma(&mut self) -> uint {
+	fn perform_vramdma(&mut self) -> u32 {
 		match self.hdma_status {
 			DMAType::NoDMA => 0,
 			DMAType::GDMA => self.perform_gdma(),
@@ -281,7 +281,7 @@ impl<'a> MMU<'a> {
 		}
 	}
 
-	fn perform_hdma(&mut self) -> uint {
+	fn perform_hdma(&mut self) -> u32 {
 		if self.gpu.may_hdma() == false || self.hdma_len == 0xFF {
 			return 0;
 		}
@@ -292,8 +292,8 @@ impl<'a> MMU<'a> {
 		return 0x10 * (if self.gbspeed == GbSpeed::Single { 4 } else { 2 });
 	}
 
-	fn perform_gdma(&mut self) -> uint {
-		let len = self.hdma_len as uint + 1;
+	fn perform_gdma(&mut self) -> u32 {
+		let len = self.hdma_len as u32 + 1;
 		for _i in range(0, len) {
 			self.perform_vramdma_row();
 		}
